Rename integration domain to unifi_gateway_refactored
