+++ custom_components/unifi_gateway_refactored/manifest.json
@@ -1,0 +1,10 @@
+{
+  "domain": "unifi_gateway_refactored",
+  "name": "UniFi Gateway Refactored",
+  "version": "0.1.0",
+  "documentation": "https://github.com/example/unifi_gateway_refactored",
+  "requirements": [],
+  "codeowners": ["@example"],
+  "config_flow": true,
+  "iot_class": "local_polling"
+}
