--- custom_components/unifi_gateway_refactory/manifest.json
@@ -1,10 +1,0 @@
-{
-  "domain": "unifi_gateway_refactory",
-  "name": "UniFi Gateway Refactory",
-  "version": "0.1.0",
-  "documentation": "https://github.com/example/unifi_gateway_refactory",
-  "requirements": [],
-  "codeowners": ["@example"],
-  "config_flow": true,
-  "iot_class": "local_polling"
-}
